refactor: separated the social list into a separate component
