+++ src/components/SocialList/SocialList.module.scss
@@ -1,0 +1,41 @@
+@import '../../styles/vars';
+
+.socials {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+
+  list-style: none;
+
+  &__item {
+    transition: opacity $transitionDuration;
+
+    &:hover {
+      opacity: $hoverOpacity;
+    }
+
+    &:active {
+      opacity: $activeOpacity;
+    }
+  }
+
+  &__icon {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    padding: 7px;
+
+    background: $brandColor;
+    border-radius: $borderRadius;
+  }
+  
+  &__link {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+
+    color: $fontDarkTheme;
+    text-decoration: none;
+  }
+}
