+++ src/pages/Contacts/Contacts.module.scss
@@ -11,7 +11,7 @@
   text-align: center;
 
   filter: drop-shadow($filterDropShadow);
-  
+
   &__text {
     margin-bottom: 20px;
     @include fontBrandStyle;
@@ -59,45 +59,5 @@
     &:hover {
       color: $fontDarkTheme;
     }
-  }
-}
-
-.socials {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-
-  list-style: none;
-
-  &__item {
-    transition: opacity $transitionDuration;
-
-    &:hover {
-      opacity: $hoverOpacity;
-    }
-
-    &:active {
-      opacity: $activeOpacity;
-    }
-  }
-
-  &__icon {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    padding: 7px;
-
-    background: $brandColor;
-    border-radius: $borderRadius;
-  }
-  
-  &__link {
-    display: flex;
-    align-items: center;
-    gap: 20px;
-
-    color: $fontDarkTheme;
-    text-decoration: none;
   }
 }
